Добавляет тест для компонента SortingOptions
