+++ src/components/main/main.jsx
@@ -6,7 +6,7 @@ import Map from "../map/map.jsx";
 import {CardsClass} from "../../const.js";
 import CitiesList from "../cities-list/cities-list.jsx";
 import SortingOptions from '../sorting-options/sorting-options.jsx';
-import {SortingTypes} from "../../const.js";
+import {sortingOffers} from "../../utils.js";
 
 const Main = (props) => {
   const {onChangeScreen, city, activeOffers} = props;
@@ -67,27 +67,6 @@ Main.propTypes = {
   onChangeScreen: PropTypes.func.isRequired,
   city: PropTypes.string.isRequired,
   activeOffers: PropTypes.array.isRequired,
-};
-
-const sortingOffers = (offers, sortType) => {
-  let sortedOffers = [];
-
-  switch (sortType) {
-    case SortingTypes.POPULAR:
-      sortedOffers = offers;
-      break;
-    case SortingTypes.PRICE_LOW_TO_HIGH:
-      sortedOffers = offers.slice().sort((a, b) => a.price - b.price);
-      break;
-    case SortingTypes.PRICE_HIGH_TO_LOW:
-      sortedOffers = offers.slice().sort((a, b) => b.price - a.price);
-      break;
-    case SortingTypes.TOP_RATED_FIRST:
-      sortedOffers = offers.slice().sort((a, b) => b.rating - a.rating);
-      break;
-  }
-
-  return sortedOffers;
 };
 
 const mapStateToProps = (state) => {
